openweave-esp32-lwip-patch-0001 : Add custom core locking
Added ability for the host environment to define locking functions that
protect entry to the LwIP stack by the LwIP tcpip thread.  This enables
coordinated use of the LwIP "raw" APIs from multiple threads.

This feature differs from LwIP's experimental TCPIP_CORE_LOCKING feature
in that it does not alter the behavior of other LwIP APIs such as the
netconn API.  Note that the two features cannot be used together.